+++ src/sentry/util/library_path.h
@@ -5,10 +5,7 @@
 
 namespace sentry::util {
 
-inline godot::String get_gdextension_library_path() {
-	godot::String library_path;
-	godot::internal::gdextension_interface_get_library_path(godot::internal::library, library_path._native_ptr());
-	return library_path;
-}
+// Returns the FS path of this shared library as it was actually loaded by the OS.
+godot::String get_loaded_gdextension_library_path();
 
 } // namespace sentry::util